% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appregistry_operations.R
\name{appregistry_list_associated_resources}
\alias{appregistry_list_associated_resources}
\title{Lists all resources that are associated with specified application}
\usage{
appregistry_list_associated_resources(
  application,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{application}{[required] The name or ID of the application.}

\item{nextToken}{The token to use to get the next page of results after a previous API
call.}

\item{maxResults}{The upper bound of the number of results to return (cannot exceed 25).
If this parameter is omitted, it defaults to 25. This value is optional.}
}
\description{
Lists all resources that are associated with specified application. Results are paginated.

See \url{https://paws-r.github.io/docs/appregistry/list_associated_resources.html} for full documentation.
}
\keyword{internal}
