% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_delete_metric_stream}
\alias{cloudwatch_delete_metric_stream}
\title{Permanently deletes the metric stream that you specify}
\usage{
cloudwatch_delete_metric_stream(Name)
}
\arguments{
\item{Name}{[required] The name of the metric stream to delete.}
}
\description{
Permanently deletes the metric stream that you specify.

See \url{https://paws-r.github.io/docs/cloudwatch/delete_metric_stream.html} for full documentation.
}
\keyword{internal}
