% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_describe_insight_rules}
\alias{cloudwatch_describe_insight_rules}
\title{Returns a list of all the Contributor Insights rules in your account}
\usage{
cloudwatch_describe_insight_rules(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{Include this value, if it was returned by the previous operation, to get
the next set of rules.}

\item{MaxResults}{The maximum number of results to return in one operation. If you omit
this parameter, the default of 500 is used.}
}
\description{
Returns a list of all the Contributor Insights rules in your account.

See \url{https://paws-r.github.io/docs/cloudwatch/describe_insight_rules.html} for full documentation.
}
\keyword{internal}
