% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatch_operations.R
\name{cloudwatch_disable_insight_rules}
\alias{cloudwatch_disable_insight_rules}
\title{Disables the specified Contributor Insights rules}
\usage{
cloudwatch_disable_insight_rules(RuleNames)
}
\arguments{
\item{RuleNames}{[required] An array of the rule names to disable. If you need to find out the names
of your rules, use
\code{\link[=cloudwatch_describe_insight_rules]{describe_insight_rules}}.}
}
\description{
Disables the specified Contributor Insights rules. When rules are disabled, they do not analyze log groups and do not incur costs.

See \url{https://paws-r.github.io/docs/cloudwatch/disable_insight_rules.html} for full documentation.
}
\keyword{internal}
