% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchrum_operations.R
\name{cloudwatchrum_list_tags_for_resource}
\alias{cloudwatchrum_list_tags_for_resource}
\title{Displays the tags associated with a CloudWatch RUM resource}
\usage{
cloudwatchrum_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource that you want to see the tags of.}
}
\description{
Displays the tags associated with a CloudWatch RUM resource.

See \url{https://paws-r.github.io/docs/cloudwatchrum/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
