% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_delete_layer}
\alias{opsworks_delete_layer}
\title{Deletes a specified layer}
\usage{
opsworks_delete_layer(LayerId)
}
\arguments{
\item{LayerId}{[required] The layer ID.}
}
\description{
Deletes a specified layer. You must first stop and then delete all associated instances or unassign registered instances. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/workinglayers-basics-delete.html}{How to Delete a Layer}.

See \url{https://paws-r.github.io/docs/opsworks/delete_layer.html} for full documentation.
}
\keyword{internal}
