% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_elastic_load_balancers}
\alias{opsworks_describe_elastic_load_balancers}
\title{Describes a stack's Elastic Load Balancing instances}
\usage{
opsworks_describe_elastic_load_balancers(StackId = NULL, LayerIds = NULL)
}
\arguments{
\item{StackId}{A stack ID. The action describes the stack's Elastic Load Balancing
instances.}

\item{LayerIds}{A list of layer IDs. The action describes the Elastic Load Balancing
instances for the specified layers.}
}
\description{
Describes a stack's Elastic Load Balancing instances.

See \url{https://paws-r.github.io/docs/opsworks/describe_elastic_load_balancers.html} for full documentation.
}
\keyword{internal}
