% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_layers}
\alias{opsworks_describe_layers}
\title{Requests a description of one or more layers in a specified stack}
\usage{
opsworks_describe_layers(StackId = NULL, LayerIds = NULL)
}
\arguments{
\item{StackId}{The stack ID.}

\item{LayerIds}{An array of layer IDs that specify the layers to be described. If you
omit this parameter, \code{\link[=opsworks_describe_layers]{describe_layers}}
returns a description of every layer in the specified stack.}
}
\description{
Requests a description of one or more layers in a specified stack.

See \url{https://paws-r.github.io/docs/opsworks/describe_layers.html} for full documentation.
}
\keyword{internal}
