% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_stacks}
\alias{opsworks_describe_stacks}
\title{Requests a description of one or more stacks}
\usage{
opsworks_describe_stacks(StackIds = NULL)
}
\arguments{
\item{StackIds}{An array of stack IDs that specify the stacks to be described. If you
omit this parameter, \code{\link[=opsworks_describe_stacks]{describe_stacks}}
returns a description of every stack.}
}
\description{
Requests a description of one or more stacks.

See \url{https://paws-r.github.io/docs/opsworks/describe_stacks.html} for full documentation.
}
\keyword{internal}
