% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_describe_volumes}
\alias{opsworks_describe_volumes}
\title{Describes an instance's Amazon EBS volumes}
\usage{
opsworks_describe_volumes(
  InstanceId = NULL,
  StackId = NULL,
  RaidArrayId = NULL,
  VolumeIds = NULL
)
}
\arguments{
\item{InstanceId}{The instance ID. If you use this parameter,
\code{\link[=opsworks_describe_volumes]{describe_volumes}} returns descriptions of
the volumes associated with the specified instance.}

\item{StackId}{A stack ID. The action describes the stack's registered Amazon EBS
volumes.}

\item{RaidArrayId}{The RAID array ID. If you use this parameter,
\code{\link[=opsworks_describe_volumes]{describe_volumes}} returns descriptions of
the volumes associated with the specified RAID array.}

\item{VolumeIds}{Am array of volume IDs. If you use this parameter,
\code{\link[=opsworks_describe_volumes]{describe_volumes}} returns descriptions of
the specified volumes. Otherwise, it returns a description of every
volume.}
}
\description{
Describes an instance's Amazon EBS volumes.

See \url{https://paws-r.github.io/docs/opsworks/describe_volumes.html} for full documentation.
}
\keyword{internal}
