% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_detach_elastic_load_balancer}
\alias{opsworks_detach_elastic_load_balancer}
\title{Detaches a specified Elastic Load Balancing instance from its layer}
\usage{
opsworks_detach_elastic_load_balancer(ElasticLoadBalancerName, LayerId)
}
\arguments{
\item{ElasticLoadBalancerName}{[required] The Elastic Load Balancing instance's name.}

\item{LayerId}{[required] The ID of the layer that the Elastic Load Balancing instance is attached
to.}
}
\description{
Detaches a specified Elastic Load Balancing instance from its layer.

See \url{https://paws-r.github.io/docs/opsworks/detach_elastic_load_balancer.html} for full documentation.
}
\keyword{internal}
