% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_get_hostname_suggestion}
\alias{opsworks_get_hostname_suggestion}
\title{Gets a generated host name for the specified layer, based on the current
host name theme}
\usage{
opsworks_get_hostname_suggestion(LayerId)
}
\arguments{
\item{LayerId}{[required] The layer ID.}
}
\description{
Gets a generated host name for the specified layer, based on the current host name theme.

See \url{https://paws-r.github.io/docs/opsworks/get_hostname_suggestion.html} for full documentation.
}
\keyword{internal}
