% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_register_rds_db_instance}
\alias{opsworks_register_rds_db_instance}
\title{Registers an Amazon RDS instance with a stack}
\usage{
opsworks_register_rds_db_instance(
  StackId,
  RdsDbInstanceArn,
  DbUser,
  DbPassword
)
}
\arguments{
\item{StackId}{[required] The stack ID.}

\item{RdsDbInstanceArn}{[required] The Amazon RDS instance's ARN.}

\item{DbUser}{[required] The database's master user name.}

\item{DbPassword}{[required] The database password.}
}
\description{
Registers an Amazon RDS instance with a stack.

See \url{https://paws-r.github.io/docs/opsworks/register_rds_db_instance.html} for full documentation.
}
\keyword{internal}
