% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_register_volume}
\alias{opsworks_register_volume}
\title{Registers an Amazon EBS volume with a specified stack}
\usage{
opsworks_register_volume(Ec2VolumeId = NULL, StackId)
}
\arguments{
\item{Ec2VolumeId}{The Amazon EBS volume ID.}

\item{StackId}{[required] The stack ID.}
}
\description{
Registers an Amazon EBS volume with a specified stack. A volume can be registered with only one stack at a time. If the volume is already registered, you must first deregister it by calling \code{\link[=opsworks_deregister_volume]{deregister_volume}}. For more information, see \href{https://docs.aws.amazon.com/opsworks/latest/userguide/resources.html}{Resource Management}.

See \url{https://paws-r.github.io/docs/opsworks/register_volume.html} for full documentation.
}
\keyword{internal}
