% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworks_operations.R
\name{opsworks_stop_stack}
\alias{opsworks_stop_stack}
\title{Stops a specified stack}
\usage{
opsworks_stop_stack(StackId)
}
\arguments{
\item{StackId}{[required] The stack ID.}
}
\description{
Stops a specified stack.

See \url{https://paws-r.github.io/docs/opsworks/stop_stack.html} for full documentation.
}
\keyword{internal}
