% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsworkscm_operations.R
\name{opsworkscm_update_server}
\alias{opsworkscm_update_server}
\title{Updates settings for a server}
\usage{
opsworkscm_update_server(
  DisableAutomatedBackup = NULL,
  BackupRetentionCount = NULL,
  ServerName,
  PreferredMaintenanceWindow = NULL,
  PreferredBackupWindow = NULL
)
}
\arguments{
\item{DisableAutomatedBackup}{Setting DisableAutomatedBackup to \code{true} disables automated or scheduled
backups. Automated backups are enabled by default.}

\item{BackupRetentionCount}{Sets the number of automated backups that you want to keep.}

\item{ServerName}{[required] The name of the server to update.}

\item{PreferredMaintenanceWindow}{}

\item{PreferredBackupWindow}{}
}
\description{
Updates settings for a server.

See \url{https://paws-r.github.io/docs/opsworkscm/update_server.html} for full documentation.
}
\keyword{internal}
