% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_describe_organization}
\alias{organizations_describe_organization}
\title{Retrieves information about the organization that the user's account
belongs to}
\usage{
organizations_describe_organization()
}
\description{
Retrieves information about the organization that the user's account belongs to.

See \url{https://paws-r.github.io/docs/organizations/describe_organization.html} for full documentation.
}
\keyword{internal}
