% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_enable_aws_service_access}
\alias{organizations_enable_aws_service_access}
\title{Enables the integration of an Amazon Web Services service (the service
that is specified by ServicePrincipal) with Organizations}
\usage{
organizations_enable_aws_service_access(ServicePrincipal)
}
\arguments{
\item{ServicePrincipal}{[required] The service principal name of the Amazon Web Services service for which
you want to enable integration with your organization. This is typically
in the form of a URL, such as \code{ service-abbreviation.amazonaws.com}.}
}
\description{
Enables the integration of an Amazon Web Services service (the service that is specified by \code{ServicePrincipal}) with Organizations. When you enable integration, you allow the specified service to create a \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/using-service-linked-roles.html}{service-linked role} in all the accounts in your organization. This allows the service to perform operations on your behalf in your organization and its accounts.

See \url{https://paws-r.github.io/docs/organizations/enable_aws_service_access.html} for full documentation.
}
\keyword{internal}
