% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations_operations.R
\name{organizations_untag_resource}
\alias{organizations_untag_resource}
\title{Removes any tags with the specified keys from the specified resource}
\usage{
organizations_untag_resource(ResourceId, TagKeys)
}
\arguments{
\item{ResourceId}{[required] The ID of the resource to remove a tag from.

You can specify any of the following taggable resources.
\itemize{
\item Amazon Web Services account – specify the account ID number.
\item Organizational unit – specify the OU ID that begins with \verb{ou-} and
looks similar to: \verb{ou-1a2b-34uvwxyz }
\item Root – specify the root ID that begins with \verb{r-} and looks similar
to: \verb{r-1a2b }
\item Policy – specify the policy ID that begins with \verb{p-} andlooks
similar to: \verb{p-12abcdefg3 }
}}

\item{TagKeys}{[required] The list of keys for tags to remove from the specified resource.}
}
\description{
Removes any tags with the specified keys from the specified resource.

See \url{https://paws-r.github.io/docs/organizations/untag_resource.html} for full documentation.
}
\keyword{internal}
