% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_delete_recommendation_template}
\alias{resiliencehub_delete_recommendation_template}
\title{Deletes a recommendation template}
\usage{
resiliencehub_delete_recommendation_template(
  clientToken = NULL,
  recommendationTemplateArn
)
}
\arguments{
\item{clientToken}{Used for an idempotency token. A client token is a unique,
case-sensitive string of up to 64 ASCII characters. You should not reuse
the same client token for other API requests.}

\item{recommendationTemplateArn}{[required] The Amazon Resource Name (ARN) for a recommendation template.}
}
\description{
Deletes a recommendation template. This is a destructive action that can't be undone.

See \url{https://paws-r.github.io/docs/resiliencehub/delete_recommendation_template.html} for full documentation.
}
\keyword{internal}
