% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiliencehub_operations.R
\name{resiliencehub_list_tags_for_resource}
\alias{resiliencehub_list_tags_for_resource}
\title{Lists the tags for your resources in your Resilience Hub applications}
\usage{
resiliencehub_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for a specific resource in your
Resilience Hub application.}
}
\description{
Lists the tags for your resources in your Resilience Hub applications.

See \url{https://paws-r.github.io/docs/resiliencehub/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
