% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_associate_product_with_portfolio}
\alias{servicecatalog_associate_product_with_portfolio}
\title{Associates the specified product with the specified portfolio}
\usage{
servicecatalog_associate_product_with_portfolio(
  AcceptLanguage = NULL,
  ProductId,
  PortfolioId,
  SourcePortfolioId = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProductId}{[required] The product identifier.}

\item{PortfolioId}{[required] The portfolio identifier.}

\item{SourcePortfolioId}{The identifier of the source portfolio.}
}
\description{
Associates the specified product with the specified portfolio.

See \url{https://paws-r.github.io/docs/servicecatalog/associate_product_with_portfolio.html} for full documentation.
}
\keyword{internal}
