% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_create_portfolio_share}
\alias{servicecatalog_create_portfolio_share}
\title{Shares the specified portfolio with the specified account or
organization node}
\usage{
servicecatalog_create_portfolio_share(
  AcceptLanguage = NULL,
  PortfolioId,
  AccountId = NULL,
  OrganizationNode = NULL,
  ShareTagOptions = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{PortfolioId}{[required] The portfolio identifier.}

\item{AccountId}{The AWS account ID. For example, \code{123456789012}.}

\item{OrganizationNode}{The organization node to whom you are going to share. If
\code{OrganizationNode} is passed in, \code{PortfolioShare} will be created for
the node an ListOrganizationPortfolioAccessd its children (when
applies), and a \code{PortfolioShareToken} will be returned in the output in
order for the administrator to monitor the status of the
\code{PortfolioShare} creation process.}

\item{ShareTagOptions}{Enables or disables \code{TagOptions } sharing when creating the portfolio
share. If this flag is not provided, TagOptions sharing is disabled.}
}
\description{
Shares the specified portfolio with the specified account or organization node. Shares to an organization node can only be created by the management account of an organization or by a delegated administrator. You can share portfolios to an organization, an organizational unit, or a specific account.

See \url{https://paws-r.github.io/docs/servicecatalog/create_portfolio_share.html} for full documentation.
}
\keyword{internal}
