% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_delete_provisioning_artifact}
\alias{servicecatalog_delete_provisioning_artifact}
\title{Deletes the specified provisioning artifact (also known as a version)
for the specified product}
\usage{
servicecatalog_delete_provisioning_artifact(
  AcceptLanguage = NULL,
  ProductId,
  ProvisioningArtifactId
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProductId}{[required] The product identifier.}

\item{ProvisioningArtifactId}{[required] The identifier of the provisioning artifact.}
}
\description{
Deletes the specified provisioning artifact (also known as a version) for the specified product.

See \url{https://paws-r.github.io/docs/servicecatalog/delete_provisioning_artifact.html} for full documentation.
}
\keyword{internal}
