% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_portfolio}
\alias{servicecatalog_describe_portfolio}
\title{Gets information about the specified portfolio}
\usage{
servicecatalog_describe_portfolio(AcceptLanguage = NULL, Id)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{[required] The portfolio identifier.}
}
\description{
Gets information about the specified portfolio.

See \url{https://paws-r.github.io/docs/servicecatalog/describe_portfolio.html} for full documentation.
}
\keyword{internal}
