% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_describe_provisioned_product}
\alias{servicecatalog_describe_provisioned_product}
\title{Gets information about the specified provisioned product}
\usage{
servicecatalog_describe_provisioned_product(
  AcceptLanguage = NULL,
  Id = NULL,
  Name = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{Id}{The provisioned product identifier. You must provide the name or ID, but
not both.

If you do not provide a name or ID, or you provide both name and ID, an
\code{InvalidParametersException} will occur.}

\item{Name}{The name of the provisioned product. You must provide the name or ID,
but not both.

If you do not provide a name or ID, or you provide both name and ID, an
\code{InvalidParametersException} will occur.}
}
\description{
Gets information about the specified provisioned product.

See \url{https://paws-r.github.io/docs/servicecatalog/describe_provisioned_product.html} for full documentation.
}
\keyword{internal}
