% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_budgets_for_resource}
\alias{servicecatalog_list_budgets_for_resource}
\title{Lists all the budgets associated to the specified resource}
\usage{
servicecatalog_list_budgets_for_resource(
  AcceptLanguage = NULL,
  ResourceId,
  PageSize = NULL,
  PageToken = NULL
)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ResourceId}{[required] The resource identifier.}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}
}
\description{
Lists all the budgets associated to the specified resource.

See \url{https://paws-r.github.io/docs/servicecatalog/list_budgets_for_resource.html} for full documentation.
}
\keyword{internal}
