% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_provisioning_artifacts}
\alias{servicecatalog_list_provisioning_artifacts}
\title{Lists all provisioning artifacts (also known as versions) for the
specified product}
\usage{
servicecatalog_list_provisioning_artifacts(AcceptLanguage = NULL, ProductId)
}
\arguments{
\item{AcceptLanguage}{The language code.
\itemize{
\item \code{en} - English (default)
\item \code{jp} - Japanese
\item \code{zh} - Chinese
}}

\item{ProductId}{[required] The product identifier.}
}
\description{
Lists all provisioning artifacts (also known as versions) for the specified product.

See \url{https://paws-r.github.io/docs/servicecatalog/list_provisioning_artifacts.html} for full documentation.
}
\keyword{internal}
