% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicecatalog_operations.R
\name{servicecatalog_list_resources_for_tag_option}
\alias{servicecatalog_list_resources_for_tag_option}
\title{Lists the resources associated with the specified TagOption}
\usage{
servicecatalog_list_resources_for_tag_option(
  TagOptionId,
  ResourceType = NULL,
  PageSize = NULL,
  PageToken = NULL
)
}
\arguments{
\item{TagOptionId}{[required] The TagOption identifier.}

\item{ResourceType}{The resource type.
\itemize{
\item \code{Portfolio}
\item \code{Product}
}}

\item{PageSize}{The maximum number of items to return with this call.}

\item{PageToken}{The page token for the next set of results. To retrieve the first set of
results, use null.}
}
\description{
Lists the resources associated with the specified TagOption.

See \url{https://paws-r.github.io/docs/servicecatalog/list_resources_for_tag_option.html} for full documentation.
}
\keyword{internal}
