% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_list_aws_default_service_quotas}
\alias{servicequotas_list_aws_default_service_quotas}
\title{Lists the default values for the quotas for the specified AWS service}
\usage{
servicequotas_list_aws_default_service_quotas(
  ServiceCode,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{ServiceCode}{[required] The service identifier.}

\item{NextToken}{The token for the next page of results.}

\item{MaxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, if any, make another call with the token returned
from this call.}
}
\description{
Lists the default values for the quotas for the specified AWS service. A default value does not reflect any quota increases.

See \url{https://paws-r.github.io/docs/servicequotas/list_aws_default_service_quotas.html} for full documentation.
}
\keyword{internal}
