% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servicequotas_operations.R
\name{servicequotas_put_service_quota_increase_request_into_template}
\alias{servicequotas_put_service_quota_increase_request_into_template}
\title{Adds a quota increase request to your quota request template}
\usage{
servicequotas_put_service_quota_increase_request_into_template(
  QuotaCode,
  ServiceCode,
  AwsRegion,
  DesiredValue
)
}
\arguments{
\item{QuotaCode}{[required] The quota identifier.}

\item{ServiceCode}{[required] The service identifier.}

\item{AwsRegion}{[required] The AWS Region.}

\item{DesiredValue}{[required] The new, increased value for the quota.}
}
\description{
Adds a quota increase request to your quota request template.

See \url{https://paws-r.github.io/docs/servicequotas/put_service_quota_increase_request_into_template.html} for full documentation.
}
\keyword{internal}
