% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_create_association_batch}
\alias{ssm_create_association_batch}
\title{Associates the specified Amazon Web Services Systems Manager document
(SSM document) with the specified managed nodes or targets}
\usage{
ssm_create_association_batch(Entries)
}
\arguments{
\item{Entries}{[required] One or more associations.}
}
\description{
Associates the specified Amazon Web Services Systems Manager document (SSM document) with the specified managed nodes or targets.

See \url{https://paws-r.github.io/docs/ssm/create_association_batch.html} for full documentation.
}
\keyword{internal}
