% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_create_ops_metadata}
\alias{ssm_create_ops_metadata}
\title{If you create a new application in Application Manager, Amazon Web
Services Systems Manager calls this API operation to specify information
about the new application, including the application type}
\usage{
ssm_create_ops_metadata(ResourceId, Metadata = NULL, Tags = NULL)
}
\arguments{
\item{ResourceId}{[required] A resource ID for a new Application Manager application.}

\item{Metadata}{Metadata for a new Application Manager application.}

\item{Tags}{Optional metadata that you assign to a resource. You can specify a
maximum of five tags for an OpsMetadata object. Tags enable you to
categorize a resource in different ways, such as by purpose, owner, or
environment. For example, you might want to tag an OpsMetadata object to
identify an environment or target Amazon Web Services Region. In this
case, you could specify the following key-value pairs:
\itemize{
\item \verb{Key=Environment,Value=Production}
\item \verb{Key=Region,Value=us-east-2}
}}
}
\description{
If you create a new application in Application Manager, Amazon Web Services Systems Manager calls this API operation to specify information about the new application, including the application type.

See \url{https://paws-r.github.io/docs/ssm/create_ops_metadata.html} for full documentation.
}
\keyword{internal}
