% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_delete_activation}
\alias{ssm_delete_activation}
\title{Deletes an activation}
\usage{
ssm_delete_activation(ActivationId)
}
\arguments{
\item{ActivationId}{[required] The ID of the activation that you want to delete.}
}
\description{
Deletes an activation. You aren't required to delete an activation. If you delete an activation, you can no longer use it to register additional managed nodes. Deleting an activation doesn't de-register managed nodes. You must manually de-register managed nodes.

See \url{https://paws-r.github.io/docs/ssm/delete_activation.html} for full documentation.
}
\keyword{internal}
