% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_association_execution_targets}
\alias{ssm_describe_association_execution_targets}
\title{Views information about a specific execution of a specific association}
\usage{
ssm_describe_association_execution_targets(
  AssociationId,
  ExecutionId,
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AssociationId}{[required] The association ID that includes the execution for which you want to
view details.}

\item{ExecutionId}{[required] The execution ID for which you want to view details.}

\item{Filters}{Filters for the request. You can specify the following filters and
values.

Status (EQUAL)

ResourceId (EQUAL)

ResourceType (EQUAL)}

\item{MaxResults}{The maximum number of items to return for this call. The call also
returns a token that you can specify in a subsequent call to get the
next set of results.}

\item{NextToken}{A token to start the list. Use this token to get the next set of
results.}
}
\description{
Views information about a specific execution of a specific association.

See \url{https://paws-r.github.io/docs/ssm/describe_association_execution_targets.html} for full documentation.
}
\keyword{internal}
