% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_describe_patch_groups}
\alias{ssm_describe_patch_groups}
\title{Lists all patch groups that have been registered with patch baselines}
\usage{
ssm_describe_patch_groups(MaxResults = NULL, Filters = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of patch groups to return (per page).}

\item{Filters}{Each element in the array is a structure containing a key-value pair.

Supported keys for \code{\link[=ssm_describe_patch_groups]{describe_patch_groups}}
include the following:
\itemize{
\item \strong{\code{NAME_PREFIX}}

Sample values: \verb{AWS-} | \verb{My-}.
\item \strong{\code{OPERATING_SYSTEM}}

Sample values: \code{AMAZON_LINUX} | \code{SUSE} | \code{WINDOWS}
}}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Lists all patch groups that have been registered with patch baselines.

See \url{https://paws-r.github.io/docs/ssm/describe_patch_groups.html} for full documentation.
}
\keyword{internal}
