% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_deployable_patch_snapshot_for_instance}
\alias{ssm_get_deployable_patch_snapshot_for_instance}
\title{Retrieves the current snapshot for the patch baseline the managed node
uses}
\usage{
ssm_get_deployable_patch_snapshot_for_instance(
  InstanceId,
  SnapshotId,
  BaselineOverride = NULL
)
}
\arguments{
\item{InstanceId}{[required] The ID of the managed node for which the appropriate patch snapshot
should be retrieved.}

\item{SnapshotId}{[required] The snapshot ID provided by the user when running
\code{AWS-RunPatchBaseline}.}

\item{BaselineOverride}{Defines the basic information about a patch baseline override.}
}
\description{
Retrieves the current snapshot for the patch baseline the managed node uses. This API is primarily used by the \code{AWS-RunPatchBaseline} Systems Manager document (SSM document).

See \url{https://paws-r.github.io/docs/ssm/get_deployable_patch_snapshot_for_instance.html} for full documentation.
}
\keyword{internal}
