% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_maintenance_window_execution}
\alias{ssm_get_maintenance_window_execution}
\title{Retrieves details about a specific a maintenance window execution}
\usage{
ssm_get_maintenance_window_execution(WindowExecutionId)
}
\arguments{
\item{WindowExecutionId}{[required] The ID of the maintenance window execution that includes the task.}
}
\description{
Retrieves details about a specific a maintenance window execution.

See \url{https://paws-r.github.io/docs/ssm/get_maintenance_window_execution.html} for full documentation.
}
\keyword{internal}
