% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_get_maintenance_window_task}
\alias{ssm_get_maintenance_window_task}
\title{Retrieves the details of a maintenance window task}
\usage{
ssm_get_maintenance_window_task(WindowId, WindowTaskId)
}
\arguments{
\item{WindowId}{[required] The maintenance window ID that includes the task to retrieve.}

\item{WindowTaskId}{[required] The maintenance window task ID to retrieve.}
}
\description{
Retrieves the details of a maintenance window task.

See \url{https://paws-r.github.io/docs/ssm/get_maintenance_window_task.html} for full documentation.
}
\keyword{internal}
