% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_resume_session}
\alias{ssm_resume_session}
\title{Reconnects a session to a managed node after it has been disconnected}
\usage{
ssm_resume_session(SessionId)
}
\arguments{
\item{SessionId}{[required] The ID of the disconnected session to resume.}
}
\description{
Reconnects a session to a managed node after it has been disconnected. Connections can be resumed for disconnected sessions, but not terminated sessions.

See \url{https://paws-r.github.io/docs/ssm/resume_session.html} for full documentation.
}
\keyword{internal}
