% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_stop_automation_execution}
\alias{ssm_stop_automation_execution}
\title{Stop an Automation that is currently running}
\usage{
ssm_stop_automation_execution(AutomationExecutionId, Type = NULL)
}
\arguments{
\item{AutomationExecutionId}{[required] The execution ID of the Automation to stop.}

\item{Type}{The stop request type. Valid types include the following: Cancel and
Complete. The default type is Cancel.}
}
\description{
Stop an Automation that is currently running.

See \url{https://paws-r.github.io/docs/ssm/stop_automation_execution.html} for full documentation.
}
\keyword{internal}
