% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssm_operations.R
\name{ssm_update_document_default_version}
\alias{ssm_update_document_default_version}
\title{Set the default version of a document}
\usage{
ssm_update_document_default_version(Name, DocumentVersion)
}
\arguments{
\item{Name}{[required] The name of a custom document that you want to set as the default
version.}

\item{DocumentVersion}{[required] The version of a custom document that you want to set as the default
version.}
}
\description{
Set the default version of a document.

See \url{https://paws-r.github.io/docs/ssm/update_document_default_version.html} for full documentation.
}
\keyword{internal}
