% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_delete_contact_channel}
\alias{ssmcontacts_delete_contact_channel}
\title{To no longer receive engagements on a contact channel, you can delete
the channel from a contact}
\usage{
ssmcontacts_delete_contact_channel(ContactChannelId)
}
\arguments{
\item{ContactChannelId}{[required] The Amazon Resource Name (ARN) of the contact channel.}
}
\description{
To no longer receive engagements on a contact channel, you can delete the channel from a contact. Deleting the contact channel removes it from the contact's engagement plan. If you delete the only contact channel for a contact, you won't be able to engage that contact during an incident.

See \url{https://paws-r.github.io/docs/ssmcontacts/delete_contact_channel.html} for full documentation.
}
\keyword{internal}
