% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_stop_engagement}
\alias{ssmcontacts_stop_engagement}
\title{Stops an engagement before it finishes the final stage of the escalation
plan or engagement plan}
\usage{
ssmcontacts_stop_engagement(EngagementId, Reason = NULL)
}
\arguments{
\item{EngagementId}{[required] The Amazon Resource Name (ARN) of the engagement.}

\item{Reason}{The reason that you're stopping the engagement.}
}
\description{
Stops an engagement before it finishes the final stage of the escalation plan or engagement plan. Further contacts aren't engaged.

See \url{https://paws-r.github.io/docs/ssmcontacts/stop_engagement.html} for full documentation.
}
\keyword{internal}
