% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_untag_resource}
\alias{ssmcontacts_untag_resource}
\title{Removes tags from the specified resource}
\usage{
ssmcontacts_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the contact or escalation plan.}

\item{TagKeys}{[required] The key of the tag that you want to remove.}
}
\description{
Removes tags from the specified resource.

See \url{https://paws-r.github.io/docs/ssmcontacts/untag_resource.html} for full documentation.
}
\keyword{internal}
