% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmcontacts_operations.R
\name{ssmcontacts_update_contact_channel}
\alias{ssmcontacts_update_contact_channel}
\title{Updates a contact's contact channel}
\usage{
ssmcontacts_update_contact_channel(
  ContactChannelId,
  Name = NULL,
  DeliveryAddress = NULL
)
}
\arguments{
\item{ContactChannelId}{[required] The Amazon Resource Name (ARN) of the contact channel you want to
update.}

\item{Name}{The name of the contact channel.}

\item{DeliveryAddress}{The details that Incident Manager uses when trying to engage the contact
channel.}
}
\description{
Updates a contact's contact channel.

See \url{https://paws-r.github.io/docs/ssmcontacts/update_contact_channel.html} for full documentation.
}
\keyword{internal}
