% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_get_incident_record}
\alias{ssmincidents_get_incident_record}
\title{Returns the details for the specified incident record}
\usage{
ssmincidents_get_incident_record(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the incident record.}
}
\description{
Returns the details for the specified incident record.

See \url{https://paws-r.github.io/docs/ssmincidents/get_incident_record.html} for full documentation.
}
\keyword{internal}
