% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_list_related_items}
\alias{ssmincidents_list_related_items}
\title{List all related items for an incident record}
\usage{
ssmincidents_list_related_items(
  incidentRecordArn,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{incidentRecordArn}{[required] The Amazon Resource Name (ARN) of the incident record containing the
listed related items.}

\item{maxResults}{The maximum number of related items per page.}

\item{nextToken}{The pagination token to continue to the next page of results.}
}
\description{
List all related items for an incident record.

See \url{https://paws-r.github.io/docs/ssmincidents/list_related_items.html} for full documentation.
}
\keyword{internal}
