% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmincidents_operations.R
\name{ssmincidents_update_deletion_protection}
\alias{ssmincidents_update_deletion_protection}
\title{Update deletion protection to either allow or deny deletion of the final
Region in a replication set}
\usage{
ssmincidents_update_deletion_protection(
  arn,
  clientToken = NULL,
  deletionProtected
)
}
\arguments{
\item{arn}{[required] The Amazon Resource Name (ARN) of the replication set you're updating.}

\item{clientToken}{A token ensuring that the operation is called only once with the
specified details.}

\item{deletionProtected}{[required] Details if deletion protection is enabled or disabled in your account.}
}
\description{
Update deletion protection to either allow or deny deletion of the final Region in a replication set.

See \url{https://paws-r.github.io/docs/ssmincidents/update_deletion_protection.html} for full documentation.
}
\keyword{internal}
