% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_describe_severity_levels}
\alias{support_describe_severity_levels}
\title{Returns the list of severity levels that you can assign to a support
case}
\usage{
support_describe_severity_levels(language = NULL)
}
\arguments{
\item{language}{The ISO 639-1 code for the language in which Amazon Web Services
provides support. Amazon Web Services Support currently supports English
("en") and Japanese ("ja"). Language parameters must be passed
explicitly for operations that take them.}
}
\description{
Returns the list of severity levels that you can assign to a support case. The severity level for a case is also a field in the CaseDetails data type that you include for a \code{\link[=support_create_case]{create_case}} request.

See \url{https://paws-r.github.io/docs/support/describe_severity_levels.html} for full documentation.
}
\keyword{internal}
