% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_operations.R
\name{support_describe_trusted_advisor_checks}
\alias{support_describe_trusted_advisor_checks}
\title{Returns information about all available Trusted Advisor checks,
including the name, ID, category, description, and metadata}
\usage{
support_describe_trusted_advisor_checks(language)
}
\arguments{
\item{language}{[required] The ISO 639-1 code for the language in which Amazon Web Services
provides support. Amazon Web Services Support currently supports English
("en") and Japanese ("ja"). Language parameters must be passed
explicitly for operations that take them.}
}
\description{
Returns information about all available Trusted Advisor checks, including the name, ID, category, description, and metadata. You must specify a language code. The Amazon Web Services Support API currently supports English ("en") and Japanese ("ja"). The response contains a TrustedAdvisorCheckDescription object for each check. You must set the Amazon Web Services Region to us-east-1.

See \url{https://paws-r.github.io/docs/support/describe_trusted_advisor_checks.html} for full documentation.
}
\keyword{internal}
