% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_delete_canary}
\alias{synthetics_delete_canary}
\title{Permanently deletes the specified canary}
\usage{
synthetics_delete_canary(Name, DeleteLambda = NULL)
}
\arguments{
\item{Name}{[required] The name of the canary that you want to delete. To find the names of
your canaries, use \code{\link[=synthetics_describe_canaries]{describe_canaries}}.}

\item{DeleteLambda}{Specifies whether to also delete the Lambda functions and layers used by
this canary. The default is false.

Type: Boolean}
}
\description{
Permanently deletes the specified canary.

See \url{https://paws-r.github.io/docs/synthetics/delete_canary.html} for full documentation.
}
\keyword{internal}
