% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_delete_group}
\alias{synthetics_delete_group}
\title{Deletes a group}
\usage{
synthetics_delete_group(GroupIdentifier)
}
\arguments{
\item{GroupIdentifier}{[required] Specifies which group to delete. You can specify the group name, the
ARN, or the group ID as the \code{GroupIdentifier}.}
}
\description{
Deletes a group. The group doesn't need to be empty to be deleted. If there are canaries in the group, they are not deleted when you delete the group.

See \url{https://paws-r.github.io/docs/synthetics/delete_group.html} for full documentation.
}
\keyword{internal}
