% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_list_tags_for_resource}
\alias{synthetics_list_tags_for_resource}
\title{Displays the tags associated with a canary or group}
\usage{
synthetics_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the canary or group that you want to view tags for.

The ARN format of a canary is
\code{arn:aws:synthetics:Region:account-id:canary:canary-name }.

The ARN format of a group is
\code{arn:aws:synthetics:Region:account-id:group:group-name }}
}
\description{
Displays the tags associated with a canary or group.

See \url{https://paws-r.github.io/docs/synthetics/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
