% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetics_operations.R
\name{synthetics_tag_resource}
\alias{synthetics_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified canary or
group}
\usage{
synthetics_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the canary or group that you're adding tags to.

The ARN format of a canary is
\code{arn:aws:synthetics:Region:account-id:canary:canary-name }.

The ARN format of a group is
\code{arn:aws:synthetics:Region:account-id:group:group-name }}

\item{Tags}{[required] The list of key-value pairs to associate with the resource.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified canary or group.

See \url{https://paws-r.github.io/docs/synthetics/tag_resource.html} for full documentation.
}
\keyword{internal}
